requirements.txt
python-frontmatter==0.3.1
